restructure proj and andd tests

+++ server/Server/Repositories/PagedResult.cs
@@ -6,7 +6,7 @@ namespace Server.Repositories
     {
         public IEnumerable<T> Items { get; set; }
         public int TotalCount { get; set; }
-        public float TotalProfit { get; set; }
+        public decimal TotalProfit { get; set; }
         public int Page { get; set; }
         public int PageSize { get; set; }
         public bool HasNext { get; set; }
